+++ src/core/HLE/D3D8/XbD3D8Types.h
@@ -494,9 +494,9 @@ typedef struct _CxbxVertexShaderStreamInfo
 {
     BOOL  NeedPatch;       // This is to know whether it's data which must be patched
 	BOOL DeclPosition;
-    DWORD HostVertexStride;
+    WORD HostVertexStride;
     DWORD NumberOfVertexElements;        // Number of the stream data types
-	DWORD CurrentStreamNumber;
+	WORD CurrentStreamNumber;
 	CxbxVertexShaderStreamElement VertexElements[32];
 }
 CxbxVertexShaderStreamInfo;
